RDKCOM-2532: RDKBDEV-667: updates following unsigned long -> uint32_t
Reason for change:
updates following unsigned long -> uint32_t data type
change in ANSC_IPV4_ADDRESS
Test Procedure: Sanity.
Risks: None.

(cherry picked from commit 8d0fb01b09afdd9ab1b38d9abcfcb0a1257d3680)

+++ source-arm/TR-181/board_sbapi/cosa_x_cisco_com_devicecontrol_apis.c
@@ -1030,7 +1030,7 @@ ANSC_STATUS
 CosaDmlDcSetWanStaticIPAddress
     (
         ANSC_HANDLE                 hContext,
-        ULONG           ipAddr
+        uint32_t                    ipAddr
     )
 {
     UNREFERENCED_PARAMETER(hContext);
@@ -1061,7 +1061,7 @@ ANSC_STATUS
 CosaDmlDcSetWanStaticSubnetMask
     (
         ANSC_HANDLE                 hContext,
-        ULONG           ipAddr
+        uint32_t                    ipAddr
     )
 {
     UNREFERENCED_PARAMETER(hContext);
@@ -1092,7 +1092,7 @@ ANSC_STATUS
 CosaDmlDcSetWanStaticGatewayIP
     (
         ANSC_HANDLE                 hContext,
-        ULONG           ipAddr
+        uint32_t                    ipAddr
     )
 {
     UNREFERENCED_PARAMETER(hContext);
@@ -1123,7 +1123,7 @@ ANSC_STATUS
 CosaDmlDcSetWanSecondIPAddr
     (
         ANSC_HANDLE                 hContext,
-        ULONG           ipAddr
+        uint32_t                    ipAddr
     )
 {
     UNREFERENCED_PARAMETER(hContext);
@@ -1149,7 +1149,7 @@ ANSC_STATUS
 CosaDmlDcSetWanBackupDefaultGateway
     (
         ANSC_HANDLE                 hContext,
-        ULONG           ipAddr
+        uint32_t                    ipAddr
     )
 {
     UNREFERENCED_PARAMETER(hContext);
@@ -1162,7 +1162,7 @@ ANSC_STATUS
 CosaDmlDcSetWanNameServer
     (
         ANSC_HANDLE                 hContext,
-        ULONG                       ipAddr,
+        uint32_t                    ipAddr,
         int                         nameServerNo
     )
 {
